fixed elo system, made elo a decimalfield, ensured two different cats each time

+++ catmash/models.py
@@ -5,6 +5,6 @@ import re
 # Create your models here.
 class pictures(models.Model):
     url = models.CharField(max_length=40)
-    rating = models.IntegerField()
+    rating = models.DecimalField(max_digits=10, decimal_places=5)
     def __unicode__(self):
         return self.url+' rating of '+str(self.rating)
